Add remap alignments functionality (#912)
* Fix issues in fine tuning

* Update configure command

* Fix error in wrong profile usage

* Update changelog for 3.3.5

+++ montreal_forced_aligner/alignment/pretrained.py
@@ -8,17 +8,22 @@ import shutil
 import time
 import typing
 from pathlib import Path
-from typing import TYPE_CHECKING, Any, Dict, Optional
 
 import sqlalchemy
 from _kalpy.matrix import DoubleMatrix, FloatMatrix
+from kalpy.aligner import KalpyAligner
 from kalpy.data import Segment
 from kalpy.utils import read_kaldi_object
 from kalpy.utterance import Utterance as KalpyUtterance
 from sqlalchemy.orm import Session, subqueryload
 
+from montreal_forced_aligner import config
 from montreal_forced_aligner.abc import TopLevelMfaWorker
-from montreal_forced_aligner.alignment.multiprocessing import AnalyzeTranscriptsFunction
+from montreal_forced_aligner.alignment.multiprocessing import (
+    AnalyzeTranscriptsFunction,
+    FineTuneArguments,
+    FineTuneFunction,
+)
 from montreal_forced_aligner.data import PhoneType, WorkflowType
 from montreal_forced_aligner.db import (
     Corpus,
@@ -27,8 +32,10 @@ from montreal_forced_aligner.db import (
     Grapheme,
     Job,
     Phone,
+    PhoneInterval,
     Speaker,
     Utterance,
+    WordInterval,
     bulk_update,
 )
 from montreal_forced_aligner.exceptions import KaldiProcessingError
@@ -39,14 +46,11 @@ from montreal_forced_aligner.helper import (
     split_phone_position,
 )
 from montreal_forced_aligner.models import AcousticModel
-from montreal_forced_aligner.online.alignment import (
-    align_utterance_online,
-    update_utterance_intervals,
-)
+from montreal_forced_aligner.online.alignment import update_utterance_intervals
 from montreal_forced_aligner.transcription.transcriber import TranscriberMixin
 from montreal_forced_aligner.utils import log_kaldi_errors, run_kaldi_function
 
-if TYPE_CHECKING:
+if typing.TYPE_CHECKING:
     from montreal_forced_aligner.abc import MetaDict
 
 __all__ = ["PretrainedAligner", "DictionaryTrainer"]
@@ -75,6 +79,8 @@ class PretrainedAligner(TranscriberMixin, TopLevelMfaWorker):
         self,
         acoustic_model_path: Path = None,
         use_reference_alignments: bool = False,
+        fine_tune: bool = False,
+        fine_tune_boundary_tolerance: typing.Optional[float] = None,
         **kwargs,
     ):
         self.acoustic_model = AcousticModel(acoustic_model_path)
@@ -82,7 +88,10 @@ class PretrainedAligner(TranscriberMixin, TopLevelMfaWorker):
         kw = self.acoustic_model.parameters
         kw.update(kwargs)
         super().__init__(**kw)
+        self.fine_tune = fine_tune
+        self.fine_tune = fine_tune_boundary_tolerance
         self.final_alignment = True
+        self.kalpy_aligner = None
 
     def setup_acoustic_model(self) -> None:
         """Set up the acoustic model"""
@@ -90,7 +99,7 @@ class PretrainedAligner(TranscriberMixin, TopLevelMfaWorker):
         os.makedirs(self.phones_dir, exist_ok=True)
         for f in ["phones.txt", "graphemes.txt"]:
             path = self.working_directory.joinpath(f)
-            if os.path.exists(path):
+            if os.path.exists(path) and not os.path.exists(os.path.join(self.phones_dir, f)):
                 os.rename(path, os.path.join(self.phones_dir, f))
         dict_info = self.acoustic_model.meta.get("dictionaries", None)
         if not dict_info:
@@ -199,7 +208,7 @@ class PretrainedAligner(TranscriberMixin, TopLevelMfaWorker):
             )
             session.commit()
 
-    def analyze_alignments(self, calculate_duration_statistics=False):
+    def analyze_alignments(self, calculate_duration_statistics=True):
         super().analyze_alignments(calculate_duration_statistics=calculate_duration_statistics)
 
     def setup(self) -> None:
@@ -210,6 +219,12 @@ class PretrainedAligner(TranscriberMixin, TopLevelMfaWorker):
             return
         begin = time.time()
         try:
+            self.initialize_database()
+            self.create_new_current_workflow(WorkflowType.alignment)
+            wf = self.current_workflow
+            if wf.done:
+                logger.info("Alignment already done, skipping.")
+                return
             os.makedirs(self.working_log_directory, exist_ok=True)
             check = self.check_previous_run()
             if check:
@@ -227,6 +242,11 @@ class PretrainedAligner(TranscriberMixin, TopLevelMfaWorker):
                 )
             self.acoustic_model.validate(self)
             self.acoustic_model.log_details()
+            self.kalpy_aligner = KalpyAligner(
+                self.acoustic_model,
+                self.lexicon_compilers,
+                **self.align_options,
+            )
 
         except Exception as e:
             if isinstance(e, KaldiProcessingError):
@@ -239,9 +259,9 @@ class PretrainedAligner(TranscriberMixin, TopLevelMfaWorker):
     @classmethod
     def parse_parameters(
         cls,
-        config_path: Optional[Path] = None,
-        args: Optional[Dict[str, Any]] = None,
-        unknown_args: Optional[typing.Iterable[str]] = None,
+        config_path: typing.Optional[Path] = None,
+        args: typing.Optional[typing.Dict[str, typing.Any]] = None,
+        unknown_args: typing.Optional[typing.Iterable[str]] = None,
     ) -> MetaDict:
         """
         Parse parameters from a config path or command-line arguments
@@ -321,18 +341,12 @@ class PretrainedAligner(TranscriberMixin, TopLevelMfaWorker):
         fmllr_trans = None
         if fmllr_string:
             fmllr_trans = read_kaldi_object(FloatMatrix, fmllr_string)
-
         text = utterance.normalized_text
         if self.use_g2p:
             text = utterance.normalized_character_text
         utterance_data = KalpyUtterance(segment, text, cmvn_string, fmllr_string)
-        ctm = align_utterance_online(
-            self.acoustic_model,
-            utterance_data,
-            self.lexicon_compilers[dictionary_id],
-            cmvn=cmvn,
-            fmllr_trans=fmllr_trans,
-            **self.align_options,
+        ctm = self.kalpy_aligner.align_utterance(
+            utterance_data, cmvn, fmllr_trans, dictionary_id=dictionary_id
         )
         update_utterance_intervals(session, utterance, ctm)
 
@@ -394,18 +408,73 @@ class PretrainedAligner(TranscriberMixin, TopLevelMfaWorker):
                 bulk_update(session, Utterance, mapping)
                 session.commit()
 
-    def align(self, workflow_name=None) -> None:
-        """Run the aligner"""
-        self.initialize_database()
-        self.create_new_current_workflow(WorkflowType.alignment, workflow_name)
-        wf = self.current_workflow
-        if wf.done:
-            logger.info("Alignment already done, skipping.")
-            return
-        self.setup()
+    def _align(self):
         if not self.use_reference_alignments:
             self.reset_manual_alignments()
+        super()._align()
+        if self.fine_tune:
+            self.fine_tune_alignments()
+
+    def align(self) -> None:
+        """Run the aligner"""
+        self.setup()
         super().align()
+
+    def fine_tune_alignments(self) -> None:
+        """
+        Fine tune aligned boundaries to millisecond precision
+        """
+        logger.info("Fine tuning alignments...")
+        all_begin = time.time()
+        with self.session() as session:
+            arguments = self.fine_tune_arguments()
+            update_mappings = []
+            for result in run_kaldi_function(
+                FineTuneFunction, arguments, total_count=self.num_utterances
+            ):
+                update_mappings.extend(result)
+            bulk_update(session, PhoneInterval, update_mappings)
+            session.flush()
+            word_update_mappings = []
+            word_intervals = (
+                session.query(
+                    WordInterval.id,
+                    sqlalchemy.func.min(PhoneInterval.begin),
+                    sqlalchemy.func.max(PhoneInterval.end),
+                )
+                .join(PhoneInterval.word_interval)
+                .group_by(WordInterval.id)
+            )
+            for wi_id, begin, end in word_intervals:
+                word_update_mappings.append({"id": wi_id, "begin": begin, "end": end})
+            bulk_update(session, WordInterval, word_update_mappings)
+            session.commit()
+        self.export_frame_shift = round(self.export_frame_shift / 10, 4)
+        logger.debug(f"Fine tuning alignments took {time.time() - all_begin:.3f} seconds")
+
+    def fine_tune_arguments(self) -> typing.List[FineTuneArguments]:
+        """
+        Generate Job arguments for :class:`~montreal_forced_aligner.alignment.multiprocessing.FineTuneFunction`
+
+        Returns
+        -------
+        list[:class:`~montreal_forced_aligner.alignment.multiprocessing.FineTuneArguments`]
+            Arguments for processing
+        """
+        args = []
+        for j in self.jobs:
+            log_path = self.working_log_directory.joinpath(f"fine_tune.{j.id}.log")
+            args.append(
+                FineTuneArguments(
+                    j.id,
+                    getattr(self, "session" if config.USE_THREADING else "db_string", ""),
+                    log_path,
+                    self.acoustic_model,
+                    self.lexicon_compilers,
+                    self.fine_tune_boundary_tolerance,
+                )
+            )
+        return args
 
 
 class DictionaryTrainer(PretrainedAligner):
